pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-ticket-triage-env"
version = "0.1.0"
description = "Ticket triage OpenEnv benchmark environment"
requires-python = ">=3.10"
dependencies = [
  "openenv-core[core]>=0.2.2",
  "openai>=1.30.0",
  "pydantic>=2.7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=4.0.0",
]

[project.scripts]
server = "ticket_triage_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["ticket_triage_env", "ticket_triage_env.server"]
package-dir = { "ticket_triage_env" = ".", "ticket_triage_env.server" = "server" }
